Fixes https://github.com/ipfs/go-bitswap/issues/306

`go test -run=XXX -benchtime=20s -bench=. -cpuprofile prof.out`

#### Creating a new message

Old:
```
         .          .    493:	// Create a new message
         .      740ms    494:	msg := bsmsg.New(false)
```

New:
```
         .          .    410:	// After processing the message, clear out its fields to save memory
         .          .    411:	defer mq.msg.Reset(false)
```

#### Adding an entry to the message

Old:
```
         .      480ms    518:		msgSize += msg.AddEntry(e.Cid, e.Priority, wantType, false)
```

New:
```
         .      260ms    528:		msgSize += mq.msg.AddEntry(e.Cid, e.Priority, wantType, false)
```

#### Getting the wantlist

Old:
```
         .      1.39s    551:		for _, e := range msg.Wantlist() {
```

Seems a little faster, but I think things may just have moved around. In any case we're calling it once now instead of twice:
New:
```
         .      220ms    417:	wantlist := message.Wantlist()
```