@Stebalien 2020-03-18
https://github.com/ipfs/go-bitswap/pull/308 may be a better solution to the CID size issues.