@Stebalien 2020-03-19
So, the race is actually in the tests. It looks like the virtual network hangs on to the message (and doesn't even bother serializing). We can probably just round-trip through a protobuf message with a bit of work.