@dirkmc 2020-03-19
It looks like the races are only showing up in the tests, I will go through and fix them