@dirkmc 2020-03-19
I updated the perf timings in the initial post at the top of this PR